clean up fargocpt loader
