+++ src/simdata/loaders/fargocpt/loadgrid.py
@@ -5,4 +5,3 @@ import os
 import numpy as np
 
 from . import loadparams
-
